pyproject.toml
[project]
name = "grpo-math"
version = "0.1.0"
description = "Compare CoT, PPO, GRPO on math reasoning with Qwen3 models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.4",
    "huggingface-hub>=1.8.0",
    "modelscope>=1.35.3",
    "torch>=2.11.0",
    "transformers>=5.4.0",
    "trl>=0.29.1",
    "wandb>=0.25.1",
    "peft>=0.15.0",
    "accelerate>=1.5.0",
    "pyyaml>=6.0",
    "tqdm>=4.67.0",
    "python-dotenv>=1.0.0",
]
